tests.hls: await all filtered-HLS writer calls
- immediately set read_once event in HLSStreamReadThread
- rename _TestSubjectHLSStreamWriter to EventedHLSStreamWriter
- move EventedHLSStreamWriter and await_write
  from test_hls_filtered to stream_hls test mixin
- add write_calls and timeout parameters to await_write
- await all writer calls in TestTwitchHLSStream
